--- public/Data/EndTimeOption.js
@@ -1,8 +1,0 @@
-const end_time_options = [
-  { value: "12:00", label: "12 : 00" },
-  { value: "13:00", label: "13 : 00" },
-  { value: "14:00", label: "14 : 00" },
-  { value: "15:00", label: "15 : 00" },
-  { value: "16:00", label: "16 : 00" },
-  { value: "17:00", label: "17 : 00" },
-];
