--- public/Data/PeopleOption.js
@@ -1,8 +1,0 @@
-const people_options = [
-  { value: "1", label: "1 명" },
-  { value: "2", label: "2 명" },
-  { value: "3", label: "3 명" },
-  { value: "4", label: "4 명" },
-  { value: "5", label: "5 명" },
-  { value: "6", label: "6 명" },
-];
