+++ src/Pages/BookMark/components/Room.jsx
@@ -3,13 +3,13 @@ import React, { Component } from "react";
 import { withRouter } from "react-router-dom";
 import "./Room.scss";
 
+const targetClass = "far fa-bookmark";
+
 class Room extends Component {
   constructor() {
     super();
     this.target = React.createRef();
-    this.state = {
-      targetClass: "far fa-bookmark",
-    };
+    this.state = {};
   }
 
   goToDetail = (e) => {
@@ -21,9 +21,18 @@ class Room extends Component {
   };
 
   render() {
-    const { info } = this.props;
+    const { info, idx } = this.props;
+    const averageRatingArr =
+      this.props.info &&
+      this.props.info.ratings.map((el) => {
+        return el.starpoint;
+      });
     return (
-      <li className="room-content-list-wrap" onClick={this.goToDetail}>
+      <li
+        key={idx}
+        className="room-content-list-wrap"
+        onClick={this.goToDetail}
+      >
         <div className="slider-content-list">
           <img src={info.img_url} />
           <div className="slider-content">
@@ -44,7 +53,13 @@ class Room extends Component {
                     <div className="stars-outer">
                       <div
                         className="stars-inner"
-                        style={{ width: `${average * 20}%` }}
+                        style={{
+                          width: `${
+                            averageRatingArr.reduce((pre, cur) => {
+                              return pre + cur / averageRatingArr.length;
+                            }, 0) * 20
+                          }%`,
+                        }}
                       ></div>
                     </div>
                     <span className="number-rating">
@@ -69,12 +84,3 @@ class Room extends Component {
 }
 
 export default withRouter(Room);
-
-const averageRatingArr =
-  this.props.info &&
-  this.props.info.ratings.map((el) => {
-    return el.starpoint;
-  });
-const average = averageRatingArr.reduce((pre, cur) => {
-  return pre + cur / averageRatingArr.length;
-}, 0);
