+++ src/config.js
@@ -3,4 +3,3 @@ export const API = "http://10.58.7.159:8000";
 
 // eslint-disable-next-line prettier/prettier
 export const YA_API = "http://192.168.200.137:8000";
-
